feat: extend test

+++ module/core/derive_tools/tests/inc/all_test.rs
@@ -7,6 +7,6 @@ use super::*;
 pub struct IsTransparent( bool );
 
 #[ derive( Debug, Clone, Copy, PartialEq, Default, TheModule::FromInner, TheModule::InnerFrom ) ]
-pub struct Age {age: usize}
+pub struct Age{ age: u32 }
 
 include!( "./only_test/all.rs" );
